add test cases for checkout feature and extract assertSnapshot method
